fix (Ex6): fix over(under)flow bug

+++ Lista-1/src/Task6.java
@@ -6,7 +6,7 @@ public class Task6 {
         Scanner scanner = new Scanner(System.in);
         int[] vectorA = new int[5];
         int[] vectorB = new int[5];
-        int[] vectorC = new int[5];
+        long[] vectorC = new long[5];
 
         for (int i = 0; i < 5; i++) {
             System.out.println("Type a number for the first vector: ");
@@ -19,8 +19,10 @@ public class Task6 {
         scanner.close();
 
         for (int i = 0; i < 5; i++) {
-            vectorC[i] = vectorA[i] + vectorB[i]; // This code is prone to over and under flow. Can you notice?
-            // We will fix it in the next commit!
+            //if we sum two very big ints, the result can be bigger than an int variable can hold.
+            //solution 1: use long to store the result of summing two ints
+            //solution 2: check if the sum is bigger and, in this case, throw a runtime exception.
+            vectorC[i] = vectorA[i] + vectorB[i];
         }
         System.out.println(Arrays.toString(vectorC));
     }
